+++ packages/orq-rc/src/orq_ai_sdk/datasets.py
@@ -12,6 +12,7 @@ class Datasets(BaseSDK):
     def list(
         self,
         *,
+        sort: Optional[models.ListDatasetsQueryParamSort] = "asc",
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
@@ -24,6 +25,7 @@ class Datasets(BaseSDK):
 
         Retrieves a paginated list of datasets for the current workspace. Results can be paginated using cursor-based pagination.
 
+        :param sort: List sorting preference.
         :param limit: A limit on the number of objects to be returned. Limit can range between 1 and 50, and the default is 10
         :param starting_after: A cursor for use in pagination. `starting_after` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, ending with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `after=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the next page of the list.
         :param ending_before: A cursor for use in pagination. `ending_before` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, starting with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `before=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the previous page of the list.
@@ -46,6 +48,7 @@ class Datasets(BaseSDK):
             base_url = self._get_url(base_url, url_variables)
 
         request = models.ListDatasetsRequest(
+            sort=sort,
             limit=limit,
             starting_after=starting_after,
             ending_before=ending_before,
@@ -116,6 +119,7 @@ class Datasets(BaseSDK):
     async def list_async(
         self,
         *,
+        sort: Optional[models.ListDatasetsQueryParamSort] = "asc",
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
@@ -128,6 +132,7 @@ class Datasets(BaseSDK):
 
         Retrieves a paginated list of datasets for the current workspace. Results can be paginated using cursor-based pagination.
 
+        :param sort: List sorting preference.
         :param limit: A limit on the number of objects to be returned. Limit can range between 1 and 50, and the default is 10
         :param starting_after: A cursor for use in pagination. `starting_after` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, ending with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `after=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the next page of the list.
         :param ending_before: A cursor for use in pagination. `ending_before` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, starting with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `before=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the previous page of the list.
@@ -150,6 +155,7 @@ class Datasets(BaseSDK):
             base_url = self._get_url(base_url, url_variables)
 
         request = models.ListDatasetsRequest(
+            sort=sort,
             limit=limit,
             starting_after=starting_after,
             ending_before=ending_before,
@@ -1061,6 +1067,7 @@ class Datasets(BaseSDK):
         self,
         *,
         dataset_id: str,
+        sort: Optional[models.ListDatasetDatapointsQueryParamSort] = "asc",
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
@@ -1074,6 +1081,7 @@ class Datasets(BaseSDK):
         Retrieves a paginated list of datapoints from a specific dataset.
 
         :param dataset_id:
+        :param sort: List sorting preference.
         :param limit: A limit on the number of objects to be returned. Limit can range between 1 and 50, and the default is 10
         :param starting_after: A cursor for use in pagination. `starting_after` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, ending with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `after=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the next page of the list.
         :param ending_before: A cursor for use in pagination. `ending_before` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, starting with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `before=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the previous page of the list.
@@ -1097,6 +1105,7 @@ class Datasets(BaseSDK):
 
         request = models.ListDatasetDatapointsRequest(
             dataset_id=dataset_id,
+            sort=sort,
             limit=limit,
             starting_after=starting_after,
             ending_before=ending_before,
@@ -1168,6 +1177,7 @@ class Datasets(BaseSDK):
         self,
         *,
         dataset_id: str,
+        sort: Optional[models.ListDatasetDatapointsQueryParamSort] = "asc",
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
@@ -1181,6 +1191,7 @@ class Datasets(BaseSDK):
         Retrieves a paginated list of datapoints from a specific dataset.
 
         :param dataset_id:
+        :param sort: List sorting preference.
         :param limit: A limit on the number of objects to be returned. Limit can range between 1 and 50, and the default is 10
         :param starting_after: A cursor for use in pagination. `starting_after` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, ending with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `after=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the next page of the list.
         :param ending_before: A cursor for use in pagination. `ending_before` is an object ID that defines your place in the list. For instance, if you make a list request and receive 20 objects, starting with `01JJ1HDHN79XAS7A01WB3HYSDB`, your subsequent call can include `before=01JJ1HDHN79XAS7A01WB3HYSDB` in order to fetch the previous page of the list.
@@ -1204,6 +1215,7 @@ class Datasets(BaseSDK):
 
         request = models.ListDatasetDatapointsRequest(
             dataset_id=dataset_id,
+            sort=sort,
             limit=limit,
             starting_after=starting_after,
             ending_before=ending_before,