+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -83,5 +83,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-10T09:56:37.933Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-10T11:10:33.958Z")
     r"""The date and time the resource was last updated"""