+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -94,5 +94,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-10T09:56:39.106Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-10T11:10:35.067Z")
     r"""The date and time the resource was created"""